use multiple content types on ILanguageClient
it turns out they already added that support, so moving to 1 language server with multiple content types and add support for live share as well on the way.
