+++ src/EditorFeatures/Text/ContentTypeNames.cs
@@ -12,5 +12,8 @@ namespace Microsoft.CodeAnalysis.Editor
         public const string XamlContentType = "XAML";
         public const string JavaScriptContentTypeName = "JavaScript";
         public const string TypeScriptContentTypeName = "TypeScript";
+
+        public const string CSharpLspContentTypeName = "C#_LSP";
+        public const string VisualBasicLspContentTypeName = "VB_LSP";
     }
 }
