+++ src/VisualStudio/CSharp/Impl/xlf/CSharpVSResources.ru.xlf
@@ -22,11 +22,6 @@
         <target state="translated">Правила форматирования C#</target>
         <note />
       </trans-unit>
-      <trans-unit id="CSharp_language_server_client">
-        <source>C# Language Server Client</source>
-        <target state="new">C# Language Server Client</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Discard">
         <source>Discard</source>
         <target state="translated">Отменить</target>
