+++ src/VisualStudio/Core/Def/ServicesVSResources.resx
@@ -1317,4 +1317,7 @@ I agree to all of the foregoing:</value>
     <value>Type Parameter</value>
     <comment>This string can be found under "Tools | Options | Text Editor | Basic | Code Style | Naming | Manage Specifications | + | Symbol kinds". All of the "NamingSpecification_VisualBasic_*" strings represent language constructs, and some of them are also actual keywords (NOT this one). Refers to the Visual Basic language concept of a "type parameter".</comment>
   </data>
+  <data name="CSharp_Visual_Basic_Language_server_client" xml:space="preserve">
+    <value>C#/Visual Basic Language Server Client</value>
+  </data>
 </root>
