+++ src/VisualStudio/Core/Test.Next/Services/LanguageServiceTests.cs
@@ -30,7 +30,7 @@ namespace Roslyn.VisualStudio.Next.UnitTests.Services
             {
                 var solution = workspace.CurrentSolution;
 
-                var results = await GetVsSearchResultsAsync(solution, WellKnownServiceHubServices.CSharpLanguageServer, "met");
+                var results = await GetVsSearchResultsAsync(solution, WellKnownServiceHubServices.LanguageServer, "met");
 
                 Assert.Equal(1, results.Count);
                 Assert.Equal(1, results[0].Symbols.Length);
@@ -55,7 +55,7 @@ namespace Roslyn.VisualStudio.Next.UnitTests.Services
             {
                 var solution = workspace.CurrentSolution;
 
-                var results = await GetVsSearchResultsAsync(solution, WellKnownServiceHubServices.CSharpLanguageServer, "met");
+                var results = await GetVsSearchResultsAsync(solution, WellKnownServiceHubServices.LanguageServer, "met");
 
                 Assert.Equal(1, results.Count);
                 Assert.Equal(4, results[0].Symbols.Length);
@@ -75,7 +75,7 @@ End Class";
             {
                 var solution = workspace.CurrentSolution;
 
-                var results = await GetVsSearchResultsAsync(solution, WellKnownServiceHubServices.VisualBasicLanguageServer, "met");
+                var results = await GetVsSearchResultsAsync(solution, WellKnownServiceHubServices.LanguageServer, "met");
 
                 Assert.Equal(1, results.Count);
                 Assert.Equal(1, results[0].Symbols.Length);
