+++ src/VisualStudio/VisualBasic/Impl/xlf/BasicVSResources.it.xlf
@@ -277,11 +277,6 @@
         <target state="translated">Convenzioni di scrittura codice VB</target>
         <note />
       </trans-unit>
-      <trans-unit id="Visual_Basic_language_server_client">
-        <source>Visual Basic Language Server Client</source>
-        <target state="new">Visual Basic Language Server Client</target>
-        <note />
-      </trans-unit>
       <trans-unit id="nothing_checking_colon">
         <source>'nothing' checking:</source>
         <target state="translated">'Controllo 'nothing':</target>
