+++ src/VisualStudio/VisualBasic/Impl/xlf/BasicVSResources.ko.xlf
@@ -277,11 +277,6 @@
         <target state="translated">VB 코딩 규칙</target>
         <note />
       </trans-unit>
-      <trans-unit id="Visual_Basic_language_server_client">
-        <source>Visual Basic Language Server Client</source>
-        <target state="new">Visual Basic Language Server Client</target>
-        <note />
-      </trans-unit>
       <trans-unit id="nothing_checking_colon">
         <source>'nothing' checking:</source>
         <target state="translated">'nothing' 검사:</target>
