+++ src/Workspaces/Remote/ServiceHub/Services/LanguageServer.cs
@@ -17,26 +17,7 @@ using LSP = Microsoft.VisualStudio.LanguageServer.Protocol;
 
 namespace Microsoft.CodeAnalysis.Remote
 {
-    // we need per language server for now since ILanguageClient
-    // doesn't allow multiple content types to be associated with
-    // one language server
-    internal class CSharpLanguageServer : LanguageServer
-    {
-        public CSharpLanguageServer(Stream stream, IServiceProvider serviceProvider)
-            : base(stream, serviceProvider, LanguageNames.CSharp)
-        {
-        }
-    }
-
-    internal class VisualBasicLanguageServer : LanguageServer
-    {
-        public VisualBasicLanguageServer(Stream stream, IServiceProvider serviceProvider)
-            : base(stream, serviceProvider, LanguageNames.VisualBasic)
-        {
-        }
-    }
-
-    internal abstract class LanguageServer : ServiceBase
+    internal sealed class LanguageServer : ServiceBase
     {
         private static readonly IImmutableSet<string> s_supportedKinds =
             ImmutableHashSet.Create(
@@ -53,13 +34,9 @@ namespace Microsoft.CodeAnalysis.Remote
                 NavigateToItemKind.Property,
                 NavigateToItemKind.Structure);
 
-        private readonly string _languageName;
-
-        public LanguageServer(Stream stream, IServiceProvider serviceProvider, string languageName)
+        public LanguageServer(Stream stream, IServiceProvider serviceProvider)
             : base(serviceProvider, stream, SpecializedCollections.EmptyEnumerable<JsonConverter>())
         {
-            _languageName = languageName;
-
             StartService();
         }
 
@@ -115,7 +92,7 @@ namespace Microsoft.CodeAnalysis.Remote
 
         private async Task SearchAsync(Solution solution, string query, int searchId, CancellationToken cancellationToken)
         {
-            var tasks = solution.Projects.Where(p => p.Language == _languageName).Select(p => SearchProjectAsync(p, cancellationToken)).ToArray();
+            var tasks = solution.Projects.Select(p => SearchProjectAsync(p, cancellationToken)).ToArray();
             await Task.WhenAll(tasks).ConfigureAwait(false);
             return;
